+++ libs/jwst/src/block/mod.rs
@@ -1,10 +1,16 @@
+mod convert;
+
 use super::{constants::sys, utils::JS_INT_RANGE, *};
 use lib0::any::Any;
 use serde::{Serialize, Serializer};
-use serde_json::Value;
+use serde_json::Value as JsonValue;
 use std::collections::HashMap;
 use yrs::{
-    types::ToJson, Array, ArrayPrelim, ArrayRef, Doc, Map, MapPrelim, MapRef, ReadTxn, Transact,
+    types::{
+        text::{Diff, YChange},
+        ToJson, Value,
+    },
+    Array, ArrayPrelim, ArrayRef, Doc, Map, MapPrelim, MapRef, ReadTxn, Text, TextPrelim, Transact,
     TransactionMut,
 };
 
@@ -28,7 +34,7 @@ impl Block {
         trx: &mut TransactionMut<'_>,
         space: &Space,
         block_id: B,
-        flavor: F,
+        flavour: F,
         operator: u64,
     ) -> JwstResult<Block>
     where
@@ -51,7 +57,7 @@ impl Block {
                 .unwrap();
 
             // init default schema
-            block.insert(trx, sys::FLAVOR, flavor.as_ref())?;
+            block.insert(trx, sys::FLAVOUR, flavour.as_ref())?;
             block.insert(trx, sys::VERSION, ArrayPrelim::from([1, 0]))?;
             block.insert(
                 trx,
@@ -206,13 +212,12 @@ impl Block {
 
     // start with a namespace
     // for example: affine:text
-    pub fn flavor<T>(&self, trx: &T) -> String
+    pub fn flavour<T>(&self, trx: &T) -> String
     where
         T: ReadTxn,
     {
         self.block
-            .get(trx, sys::FLAVOR)
-            .or_else(|| self.block.get(trx, sys::FLAVOUR))
+            .get(trx, sys::FLAVOUR)
             .unwrap_or_default()
             .to_string(trx)
     }
@@ -458,99 +463,6 @@ impl Block {
             .iter(trx)
             .position(|c| c.to_string(trx) == block_id)
     }
-
-    pub fn to_markdown<T>(&self, trx: &T, state: &mut MarkdownState) -> Option<String>
-    where
-        T: ReadTxn,
-    {
-        match self.get(trx, "text").map(|t| t.to_string()) {
-            Some(text) => match self.flavor(trx).as_str() {
-                "affine:code" => {
-                    state.numbered_count = 0;
-                    match self.get(trx, "language").map(|v| v.to_string()).as_deref() {
-                        Some(language) => Some(format!("``` {}\n{}\n```\n", language, text)),
-                        None => Some(format!("```\n{}\n```\n", text)),
-                    }
-                }
-                format @ "affine:paragraph" => {
-                    state.numbered_count = 0;
-                    match self.get(trx, "type").map(|v| v.to_string()).as_deref() {
-                        Some(
-                            head @ "h1" | head @ "h2" | head @ "h3" | head @ "h4" | head @ "h5",
-                        ) => Some(format!(
-                            "{} {}\n",
-                            "#".repeat(head[1..].parse().unwrap()),
-                            text
-                        )),
-                        Some("quote") => Some(format!("> {text}\n")),
-                        Some("text") => Some(format!("{text}\n")),
-                        r#type @ Some(_) | r#type @ None => {
-                            if let Some(r#type) = r#type {
-                                warn!("Unprocessed format: {format}, {}", r#type);
-                            } else {
-                                warn!("Unprocessed format: {format}");
-                            }
-                            Some(text)
-                        }
-                    }
-                }
-                format @ "affine:list" => {
-                    match self.get(trx, "type").map(|v| v.to_string()).as_deref() {
-                        Some("numbered") => {
-                            state.numbered_count += 1;
-                            Some(format!("{}. {text}\n", state.numbered_count))
-                        }
-                        Some("todo") => {
-                            state.numbered_count += 1;
-                            let clicked = self
-                                .get(trx, "checked")
-                                .map(|v| v.to_string() == "true")
-                                .unwrap_or(false);
-                            Some(format!("[{}] {text}\n", if clicked { "x" } else { " " }))
-                        }
-                        Some("bulleted") => {
-                            state.numbered_count += 1;
-                            Some(format!("- {text}\n"))
-                        }
-                        r#type @ Some("text") | r#type @ Some(_) | r#type @ None => {
-                            state.numbered_count = 0;
-                            if let Some(r#type) = r#type {
-                                warn!("Unprocessed format: {format}, {}", r#type);
-                            } else {
-                                warn!("Unprocessed format: {format}");
-                            }
-                            Some(text)
-                        }
-                    }
-                }
-                format => {
-                    state.numbered_count = 0;
-                    warn!("Unprocessed format: {format}");
-                    Some(text)
-                }
-            },
-            None => match self.flavor(trx).as_str() {
-                "affine:divider" => {
-                    state.numbered_count = 0;
-                    Some("---\n".into())
-                }
-                "affine:embed" => {
-                    state.numbered_count = 0;
-                    match self.get(trx, "type").map(|v| v.to_string()).as_deref() {
-                        Some("image") => self
-                            .get(trx, "sourceId")
-                            .map(|v| format!("![](/api/workspace/{}/blob/{})\n", self.id, v)),
-                        _ => None,
-                    }
-                }
-                format => {
-                    state.numbered_count = 0;
-                    warn!("Unprocessed format: {format}");
-                    None
-                }
-            },
-        }
-    }
 }
 
 #[derive(Default)]
@@ -568,12 +480,15 @@ impl Serialize for Block {
 
         let mut buffer = String::new();
         any.to_json(&mut buffer);
-        let any: Value = serde_json::from_str(&buffer).unwrap();
+        let any: JsonValue = serde_json::from_str(&buffer).unwrap();
 
         let mut block = any.as_object().unwrap().clone();
-        block.insert("sys:id".to_string(), Value::String(self.block_id.clone()));
+        block.insert(
+            constants::sys::ID.to_string(),
+            JsonValue::String(self.block_id.clone()),
+        );
 
-        Value::Object(block).serialize(serializer)
+        JsonValue::Object(block).serialize(serializer)
     }
 }
 
@@ -595,7 +510,7 @@ mod test {
             assert_eq!(block.id, "workspace");
             assert_eq!(block.space_id, "space");
             assert_eq!(block.block_id(), "test");
-            assert_eq!(block.flavor(&t.trx), "affine:text");
+            assert_eq!(block.flavour(&t.trx), "affine:text");
             assert_eq!(block.version(&t.trx), [1, 0]);
         });
 
@@ -605,7 +520,7 @@ mod test {
 
             let block = space.get(&t.trx, "test").unwrap();
 
-            assert_eq!(block.flavor(&t.trx), "affine:text");
+            assert_eq!(block.flavour(&t.trx), "affine:text");
             assert_eq!(block.version(&t.trx), [1, 0]);
         });
     }